requirements-dev.txt
bandit
hacking==1.1.0
Legobot==1.2.4
pytest
pyyaml